refactor(Questions): mainView,lineView...->view
